requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.0.1
python-dotenv==0.19.0
PyPDF2==3.0.1
numpy==1.23.5
pandas==1.5.3
matplotlib==3.5.3
seaborn==0.12.2
openpyxl==3.0.9 